[#245] refactor: import문 묶기, 주석 제거, 사용되지 않는 import 제거
