+++ pages/profile/[profileId]/favorite.tsx
@@ -1,10 +1,8 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import PageLayout from "@/components/common/pageLayout";
-import UserInfo from "@/components/common/userInfo";
 import React, { useEffect, useState } from "react";
 import { useRouter } from "next/router";
+import { PageLayout, UserInfo, OtherFilterMenu } from "@/components/common";
 import OtherBookmark from "@/components/otherBookmark/otherBookmarkTemplate";
-import OtherFilterMenu from "@/components/common/filterMenu/otherFilterMenu";
 import profileAPI from "@/utils/apis/profile";
 import { ProfileDetail } from "@/types/model";
